chore: remove testing related commands
We were not using these, and it also removes some code that would not compile on Windows.

There is also another little piece of code for setting file permissions on `safenode_rpc_client`
which would not compile on Windows. I'm not sure we actually need to use the RPC client any more,
but rather than remove the code, I just used conditional compilation.