feat: provide vm size args for `deploy` cmd
The node size will now be variable on the command line.

Upscaling is not supported here. We would need to extend the inventory to keep track of the size
overrides, such that the size for the initial deploy would be supplied for the upscale. Otherwise
Terraform would possibly try and delete the VMs and deploy new ones with a smaller size.