feat: replace safe bin with autonomi
The `sn_releases` reference is updated to the latest version, which replaces the `ReleaseType::Safe`
variant with `ReleaseType::Autonomi`, which will ensure the correct version of the client is
obtained.

The commit also removes other traces of the faucet that are no longer necessary. The
`ReleaseType::Faucet` variant was removed from the latest version of `sn_releases`.