feat: capture all listener addresses in inventory
Now that we have websockets enabled, there is an interest in finding listener addresses that are not
using QUIC.

We now store all the listener addresses, and we can filter them when reporting. The `inventory`
command adds an optional `full` flag that will print the whole list, grouped by QUIC and websockets.

For some reason there was a filter applied to the printing of peers with respect to a bootstrap
deployment. That was removed. I can't think why I added it in the first place; it would be useful to
print the peers even if the deployment is a bootstrap type.