feat: use `network-contacts` for prod environment
If we are using a production network, the autonomi client should not provide an explicit peer and
instead use the network contacts. In production it's possible for the single peer to become
unavailable and this would then break the uploader.

We need to use a template to do this, so the file was moved to a jinja template.