feat: differentiate 'custom' and anvil networks
Previously we were using the 'custom' network type for doing a deployment with Anvil, but what we
actually need the custom case for is when we want to use a different token contracts address. This
will still use Arbitrum Sepolia, it's just that it will use a different token address. However, the
payment address and the RPC URL still need to be supplied with the command. This is a closer match
to how the `custom` network type is used in the node manager.

As part of this commit, the counts for the Ansible banners were removed. They are not very useful
and were difficult to maintain based on all the different options.