chore: change various telegraf configuration details
In the base images, Telegraf has been disabled from automatically starting, because we don't want it
to start transmitting new metric names to the production database. Various lines for starting the
service have been commented out. These can be enabled again when we have the setup working.

Environment variables related to Telegraf are now prefixed with `ANTNODE_` or `ANT_` rather than
safe-based prefixes.

There's also a change in here to get `antcltd` to deploy correctly when a version-based deployment
is used.