chore: apply `telegraf-configuration` role before `node` role
It is sometimes possible that the application of the `node` role will result in an error, such as
some nodes failing to start. In this case, Ansible will stop applying remaining roles on hosts where
the failure occurred. For this reason, we want to configure and start Telegraf before we attempt to
start the nodes, so that we will still collect monitoring data even if one or two nodes fail to
start.