feat: provide `reset-to-n-nodes` command
This command runs a playbook that uses a Bash script to clear out existing nodes then create the
specified number of new nodes.

It is intended to be used in the production environment to reset nodes for a particular environment,
to avoid having to bootstrap a new one.

In production, there will be intervals applied to the `stop` and `start` commands such that the
existing node services will be stopped gradually, and the new services will come online slowly,
probably with an interval of about 5 minutes between each.