feat: handle missing or altered genesis node
In the case of our current production environment, I accidentally reset the node registry on the
genesis VM, which meant that the inventory for the production environment could no longer be
generated.

I since started 5 nodes on the genesis VM, so I'm just going to take the first node in the registry
even if there are none with the `--first` flag.

It could also be possible that the original genesis VM was destroyed. So the inventory process
treats it as optional.