refactor: use shared mechanism for apply and plan args
The `terraform plan` and `terraform apply` arguments should be identical to each other, such that
the plan will be producing the same change set as the apply. There were different code paths for
each of them, which meant they could diverge easily.

Extract a full function for obtaining the resource value from the Terraform state. The closure was
large, so it was making the enclosing function more difficult to read. The `vm_size` variable was
renamed because it was misleading: the function retrieves the value of any resource, not just the
size of a VM.

BREAKING CHANGE: I removed the `plan` command as part of this refactor. This command on its own just
didn't make a whole lot of sense. It makes more sense to be used with an upscale.