feat: obtain machine sizes from previous deploy
It's possible to launch the original deployment with values different from those in the tfvars file.
If you then try to upscale, Terraform will then want to apply the values from the tvfars file, which
could be smaller or larger.

The sizes are therefore obtained from the original deployment by inspecting the current state.