fix: peer cache node provisioning
We used to have a separate function for provisioning the peer cache nodes and this invoked the
correct playbook. However, as part of the NAT rework, this function was removed in favour of the
`provision_nodes` function that shared most of the code. The only thing was, when this refactor took
place, it meant the nodes playbook was being specified when the peer cache nodes were being
provisioned.

This has now been fixed to select the playbook based on the node type.