feat: provide `--to-genesis` arg on `deploy` cmd
A first step in breaking the deploy process down such that different parts could run in parallel.

Using this argument runs the deployment to the point where the genesis node is running. This should
allow the peer cache and generic nodes to then be provisioned in parallel, perhaps in a GHA
workflow.