chore: vary base mount point for striped volume
I want to use this role to mount a volume on the uploaders too, so the base mount point may be
different.

I renamed the role to something a bit more specific that I think captures the purpose.

I've also taken out a role being applied to the playbook that formatted a disk if we were using AWS
as the provider. We will hopefully move to AWS soon, but for the time being this was causing a bit
of confusion and I don't think it will be used; we'll probably need the striped volume setup for AWS
too.