chore: remove logstash setup
All references to logstash is removed in the Rust, Ansible and Terraform code.

At some point a long time ago we used Logstash for log forwarding, but it's completely redundant now
that we are using Telegraf to forward the logs to ELK. All it was doing was adding noise to the
codebase.

I also removed Ansible roles related to Filebeat, because this is something we are not using any
more, and there were references to Logstash here.