feat: provide `uploaders deploy` command
Creates stand-alone uploaders that are not part of another environment or whole deployment.

Right now the name of the environment needs to be different from the existing environment, but we
will probably further break down the deployment such that each of the units are separate.

A new Terraform manifest is supplied that only defines the infrastructure for the uploaders, but on
the Ansible side, almost nothing changed. The only modification was for the `build` playbook: it was
changed to accommodate building single binaries, since the uploaders only require the `ant` binary.

The ability to supply pre-funded wallet keys is available, since we're going to be testing against
the production environment. We can also specify a maximum number of uploads, which again will be
useful for the production network, to avoid burning through funds.