feat: support deployments in multiple regions
We now have a requirement to deploy a network in another region, so the deployment-related commands
are modified accordingly to provide a `--region` argument.

Droplet images apply per region, so a set needed to be generated for `ams3`. This obviously produces
a different set of image IDs, so those are put in tfvars files which will be applied depending on
which region is specified.