feat: enable downloaders by default
The `--enable-downloaders` flag is removed and so the downloaders are now provided by default.

Each can be disabled individually using three `--disable-` arguments corresponding to each service.

This is mainly to support a scenario for testing in the production environment, where we will only
want to download one file with a specific address. In this case, the downloader will be operating in
'performance' mode.