feat: support downloading specific file
In performance mode we now support downloading a file at a specific address. This will be used for
testing download performance on the production environment. When the `--file-address` argument is
used on the `client deployment` command, the downloader will be setup to download this file, then go
into the performance mode loop.

There are several other changes in this commit that supported this change:
* Use named arguments in downloader and uploader scripts.
* Make `--peer` and `--network-contacts-url` args optional. This is for easier usage for client
  deployments that test against production. The named arguments for the scripts facilitated this
  change.
* Support multiple regions for client deployment.
* Rename some variables in uploader/downloader setups to make them more generic, e.g.,
  `genesis_multiaddr` -> `peer`.
* Provide opportunity to disable uploaders. In some performance tests for the production
  environment, we wouldn't need any uploaders to be running. However, the uploader role still needs
  to be applied because it sets up the context for the `antX` users, and those are required for the
  downloaders.