chore: change `--network-id` from optional to required
On the `deploy` and `client deploy` commands, the network ID is now mandatory.

This is because we removed the `--testnet` argument from all the binaries. We now also have an
`--alpha` argument on the ant binaries that will choose a network ID of 2, which is reserved for our
new alpha network. So we could also add an `--alpha` argument and keep the `--network-id` optional,
but since we are just using an internal tool here, it's easy enough to just assert that the network
ID should always be explicitly supplied.

If we want to do a production deployment, we will set it to 1, and if we want an alpha deploy, we
set it to 2. Anything else will be considered a development network of some kind.