feat: downloaders log to external disk
The new CLN branch seems to have much more verbose logging, which caused the disks to fill up on the
client machines, since the downloaders were never setup to log to the alternate locations.

We can address the logging output, but in principle it's just good for the logs for the downloaders
to also go to the external disk.