chore: read auth token direct from manifest
The inventory plugin in the `community.digitalocean` collection says it supports reading several
environment variables, but it *does not* work when you have them set. I changed the template to read
the token directly from the environment variable and confirmed this to be working.

This means we can switch to the latest version of Ansible and the plugin, which may solve the
problem we have with deploying multiple environments.