chore: full cone node and gateway hosts use same size
The previous setup was for the full cone host to share the same size as the generic hosts. Not
exactly sure what the reason was for that, but in any case, it's now been changed such that the full
cone host can vary its size independently of the generic host, and the corresponding gateway VM will
also share the same size.