feat: provide `static-downloaders` ansible role
This is similar to the `downloaders` role, but it is self contained and downloads from a static list
of files rather than files that have been uploaded in the same testnet session.

It has the same three services as the `downloaders` role and writes out metrics CSV files with the
same data so they can be read by existing dashboards.

Two of the services operate slightly differently. The 'delayed verifier' iterates through the whole
list of files, attempting to download them all with replicated copies. By default the 'performance
verifier' is downloading the first file in the CSV, for which a larger, 4GB file has been
deliberately selected.

All the services run with a batch size of 16. They use separate variables so that later on they can
be customised to have different values if we want them to.

The role is intended to be used as a continuous downloader for the production environment.