feat: provide `clients deploy-static-downloaders` cmd
This is similar to the `clients deploy` command.

It uses the same Terraform setup, but a new Ansible playbook is used that applies the
`static-downloaders` role.

Until the downloaders implement streaming, a machine with large memory needs to be used. This is
because the three downloaders could be downloading a large file at the same time. If we didn't do
this the downloads would be subject to random crashing.