feat: sleep in static downloader to reduce bandwidth
Each service will pause for 15 minutes approximately once every hour.

In the case of the random or performance verifiers, those will sleep once an hour based on an
elapsed time. The delayed verifier will stop for 15 minutes after it's been through a cycle of each
file, which should roughly be once an hour.

Also fix up a new Clippy warning.