feat: provide `upnp_emulation` ansible role/playbook
The role sets up the `miniupnpd` server and other configuration based on Shu's research.

The playbook applies this role, then the `node` role, the latter of which has minor modifications to
conditionally control the use of `--no-upnp` and `--node-ip`.

From what I can observe, the nodes running on the UPnP hosts seem to be working OK without using
`--no-upnp`.