feat: support directories and non-archive files
The static downloaders are extended to support different kinds of files, namely those without
metadata archives, and directories.

For validating directories, 'manifests' are provided with hashes of all the files and a utility
script is included for generating a manifest; this can be used if we decide to add other
directories.

The test directories provided are taken from a dataset I have in my personal possession that I wish
to make publicly available on the network. In terms of size, they are 'small', 'medium' and 'large',
with quite a lot of files and subdirectories.

With the introduction of multiple file types, the service Bash script was becoming convoluted, so it
was converted to Python. Along with this conversion, an extra feature was added to detect service
crashes, which happened in some of our testing. Prior to this, if the service crashed it did not
register a failure in the metrics. The service maintains some simple state it uses during start up
to determine if the previous run resulted in a crash, in which case it registers a `SERVICE_CRASH`
failure.

Using Python probably also means the other downloader role/service could soon be merged together
with this one, since it's much easier to coordinate different scenarios in Python than it is with
Bash.