feat: provide `deploy-static-uploader` command
The command provides a client machine with a service that will upload static content. The uploads
will go through the metrics system. This is mainly intended for use with the production network, but
it could potentially be used for testnets. We can use this for uploading non-junk data to the
production network as a test.

The service looks in a directory and will upload any content it finds as public data. Either files or
directories can be placed in the directory. For files, it will upload them without archives. When
all the content is processed the service will stop.

The idea is the user can place any files to upload in this directory, then start the service. The
user can take a copy of the service logs to get the addresses of all the content uploaded.