chore: handle `__ansible_unsafe` wrapper in json
BREAKING CHANGE

In the newer version of Ansible, the `ansible-inventory` binary produces different JSON output.
Values previously available as strings are now wrapped inside a dictionary with an
`__ansible_unsafe` key. I can see Github issues relating to this from last year, but it has only
surfaced for me now when I updated to the latest version.

This change was breaking our inventory parsing. It has now been modified to accommodate the new
output. We all need to be on the latest version of Ansible to use it.