fix: obtain antnode version without assuming specific service
The current mechanism for obtaining the `antnode` version was assuming the existence of an
`antnode1` service. In the case of the current `PROD-02` environment, with respect to the peer cache
nodes, some of the services had been deleted in a special upgrade process.

This new mechanism uses the `status` command to obtain the running services and gets the version
information from one of them.