chore: remove publishing from release
I want to merge #341 and release it so that I don't have to wait on it compiling for the test runs.

However, the branch currently has a Git-based crate reference, which can't be changed until that
branch is released, which won't be for a while. Therefore, I'm going to disable publishing for this
crate. It really doesn't need to be published anyway.

The release process will just upload the binary to S3 and tag the repository with the version.