chore: add `rustfmt` to build components
For some reason, in the build job I have encountered an error related to the `ant-protocol` crate
about a missing `rustfmt` component. This is despite the fact that there doesn't appear to be a new
version of Rust.

We also got this error on the main `autonomi` repo, but it appears Roland has already made the
relevant change on his branch. So as a workaround, I'm going to just add the component in the
workflow.