feat: use service number to select address
When multiple chunk tracking services have been defined and there have been a small number of
uploads, it's quite probable the same address could be selected for more than one service. We can
use the service number as the row index into the metrics file with the uploads, and since each row
has a unique address, we will get a different address selected.