chore: split files into categories
For the data retrieval service, the previous `prod.csv` is split into several files, roughly
categorizing the files into large and small.

This is so quicker data retrieval runs can be performed to check for smaller files. The process
takes significantly longer when larger files are involved.

The service definition is also changed to only restart on failures, so it will stop at the end of a
run.