feat: provide `deploy-symlinked-antnodes` cmd
This command provides a deployment of a single droplet that is setup to run multiple nodes using a
symlinked binary.

It's being kept separate from the main testnet deployment because it will only be used in certain
kinds of testing scenarios. We happen to know that this kind of setup is in use by large node
operators.